% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hanlp.R
\name{hanlp.convert}
\alias{hanlp.convert}
\title{Convert a chinese sentence.}
\usage{
hanlp.convert(text, mode = "t2s")
}
\arguments{
\item{text}{a chinese sentence .}

\item{mode}{transformation mode, default \code{t2s}.}
}
\value{
a string.
}
\description{
The transformation of traditional Chinese and simplified Chinese
}
\details{
\describe{
\item{\code{s2t}}{ : Simplified Chinese to Traditional Chinese}
\item{\code{t2s}}{ : Traditional Chinese to Simplified Chinese}
\item{\code{s2tw}}{ : Simplified Chinese to Traditional Chinese (Taiwan Standard)}
\item{\code{tw2s}}{ : Traditional Chinese (Taiwan Standard) to Simplified Chinese}
\item{\code{s2hk}}{ : Simplified Chinese to Traditional Chinese (Hong Kong Standard)}
\item{\code{hk2s}}{ : Traditional Chinese (Hong Kong Standard) to Simplified Chinese}
\item{\code{s2tw}}{ : Simplified Chinese to Traditional Chinese (Taiwan Standard)}
\item{\code{tw2s}}{ : Traditional Chinese (Taiwan Standard) to Simplified Chinese}
\item{\code{t2tw}}{ : Traditional Chinese  to Taiwan Standard}
\item{\code{t2hk}}{ : Traditional Chinese  to Hong Kong Standard}
\item{\code{tw2hk}}{ : Taiwan Standard   to Hong Kong Standard}
\item{\code{hk2hw}}{ : Hong Kong Standard to Taiwan Standard}
\item{\code{pinyin}}{ : to Chinese pinyin} }
}
\examples{
\dontrun{
hanlp.convert('hello world')
hanlp.convert('hello world',mode='s2t')
}
}
\author{
\link{https://github.com/qxde01/RHanLP}
}
