% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hanlp.R
\name{hanlp.extractPhrase}
\alias{hanlp.extractPhrase}
\title{Extract key Phrase.}
\usage{
hanlp.extractPhrase(text, size = 5L)
}
\arguments{
\item{text}{A Chinese sentence in UTF-8 .}

\item{size}{the number of key Phrase.}
}
\value{
a vector of Phrase  which have been extracted.
}
\description{
A function  extract key Phrase from text .
}
\examples{
\dontrun{
hanlp.extractPhrase("hello world!")
}
}
\author{
\link{https://github.com/qxde01/RHanLP}
}
