% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hanlp.R
\name{hanlp.extractSummary}
\alias{hanlp.extractSummary}
\title{Extract key sentences.}
\usage{
hanlp.extractSummary(text, size = 5L)
}
\arguments{
\item{text}{A Chinese sentences in UTF-8 .}

\item{size}{the number of key sentences.}
}
\value{
a vector of sentences  which have been extracted.
}
\description{
A function  extract key sentences from text ,textRank algorithm .
}
\examples{
\dontrun{
hanlp.extractSummary("hello world!")
}
}
\author{
\link{https://github.com/qxde01/RHanLP}
}
