% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hanlp.R
\name{hanlp.extractWords}
\alias{hanlp.extractWords}
\title{Extract key words.}
\usage{
hanlp.extractWords(text = "", size = 10)
}
\arguments{
\item{text}{A Chinese sentence in UTF-8 .}

\item{size}{the number of key words.}
}
\value{
a vector of words  which have been extracted.
}
\description{
A function  extract key words from text ,textRank algorithm .
}
\examples{
\dontrun{
hanlp.extractWords("hello world!")
}
}
\author{
\link{https://github.com/qxde01/RHanLP}
}
