% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hanlp.R
\name{hanlp.loadDict}
\alias{hanlp.loadDict}
\title{Load user's dict into  dictionary.}
\usage{
hanlp.loadDict(dict_path = NULL)
}
\arguments{
\item{dict_path}{user's dict file path}
}
\description{
A function load user's dict into  dictionary .
}
\examples{
\dontrun{
hanlp.loadDict(dict_path='dict.txt')
}
}
\author{
\link{https://github.com/qxde01/RHanLP}
}
