% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hanlp.R
\docType{data}
\name{hanlp.naiveBayesClassify}
\alias{hanlp.naiveBayesClassify}
\title{Naive Bayes classifier.}
\format{\code{\link{R6Class}} object.}
\usage{
hanlp.naiveBayesClassify
}
\description{
a R6 class of naive Bayes classifier.
}
\details{
A R6class of naive Bayes classifier .
}
\section{Usage}{

For usage details see \bold{Methods, Arguments and Examples} sections.
\preformatted{
naiveBayes = hanlp.naiveBayesClassify$new()
naiveBayes$train(file_folder)
naiveBayes$predict(text)
naiveBayes$test(test_data)
naiveBayes$getModel()
}
}

\section{Methods}{

\describe{
  \item{\code{$new()}}{Constructor for Naive Bayes classifier.}
  \item{\code{$train(file_folder)}}{Train Naive Bayes classifier,detail in  https://github.com/hankcs/HanLP/wiki .}
  \item{\code{$predict(text)}}{Predict \code{text} category.}
  \item{\code{$test(test_data)}}{Predict  a batch of text categories,\code{test_data} is a character vector.}
  \item{\code{$getModel()}}{Output some infomation of Naive Bayes model. }
}
}

\examples{
\dontrun{
naiveBayes = hanlp.naiveBayesClassify$new()
naiveBayes$train(file_folder)
naiveBayes$predict(text)
naiveBayes$test(test_data)
naiveBayes$getModel()
}

}
\author{
qxde01
}
\keyword{datasets}
