% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newWordDiscover.R
\name{hanlp.newWordDiscover}
\alias{hanlp.newWordDiscover}
\title{new word discovering .}
\usage{
hanlp.newWordDiscover(doc, size = 100L, max_len = 5L, min_freq = 5e-05,
  min_entropy = 0.4, min_aggregation = 1.2, filter = TRUE)
}
\arguments{
\item{doc}{A big Chinese document in UTF-8 .}

\item{size}{output word number ,default 100.}

\item{max_len}{max word length default 5.}

\item{min_freq}{min word freq ,default 0.00005.}

\item{min_entropy}{min word entropy,default 0.4.}

\item{min_aggregation}{min mutual information of words,default 1.2.}

\item{filter}{Whether or not to filter out the words already existing in the HanLP dict,default \code{TRUE}.}
}
\value{
a character .
}
\description{
A function of new word discovering .
}
\examples{
\dontrun{
hanlp.newWordDiscover("hello world!")
}
}
\author{
qxde01
}
