% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hanlp.R
\name{hanlp.parseDependency}
\alias{hanlp.parseDependency}
\title{dependency parsing.}
\usage{
hanlp.parseDependency(text, mode = "NN")
}
\arguments{
\item{text}{A Chinese sentence in UTF-8 .}

\item{mode}{\code{NN,CRF,MaxEnt},default \code{NN} mode. \code{NN} is dependency parsing,\code{CRF} and \code{MaxEnt} are Semantic Dependency Parsing.}
}
\value{
a data.frame.
}
\description{
A function of dependency parsing .
}
\examples{
\dontrun{
hanlp.parseDependency("hello world!")
}
}
\author{
\link{https://github.com/qxde01/RHanLP}
}
