% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hanlp.R
\name{hanlp.segment}
\alias{hanlp.segment}
\title{Sengment a sentence.}
\usage{
hanlp.segment(text, nature = TRUE, mode = "standard")
}
\arguments{
\item{text}{A Chinese sentence in UTF-8 .}

\item{nature}{Whether to recognise the nature of the words.}

\item{mode}{segment mode,default \code{standard} mode,other mode contains \code{fast,NER,index,TCS,CRF,NShort,HMM}}
}
\value{
a vector of words  which have been segmented.
}
\description{
A function segment Chinese sentence into words.
}
\examples{
\dontrun{
hanlp.segment("hello world!")
}
}
\author{
\link{https://github.com/qxde01/RHanLP}
}
