% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hanlp.R
\docType{data}
\name{hanlp.suggest}
\alias{hanlp.suggest}
\title{text suggester.}
\format{\code{\link{R6Class}} object.}
\usage{
hanlp.suggest
}
\description{
a R6 class of text suggester.
}
\details{
A function of text suggest .
}
\section{Methods}{

\describe{
  \item{\code{$Suggester}}{a java object.}
  \item{\code{$addSentence(x)}}{add sentences into\code{Suggester},\code{x} is a character vector.}
  \item{\code{$suggest(word="",size=5L)}}{suggest \code{size} sentences for a \code{word} }
}
}

\examples{
\dontrun{
sug<-hanlp.suggest$new()
sug$addSentence(c('hello world','My R world'))
sug$suggest(word="world",size=2L)
}

}
\author{
\link{https://github.com/qxde01/RHanLP}
}
\keyword{datasets}
