% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hanlp.R
\name{hanlp.updateWord}
\alias{hanlp.updateWord}
\title{Add,remove,get a word.}
\usage{
hanlp.updateWord(x, mode = "insert")
}
\arguments{
\item{x}{a word like \code{c('word')} or  \code{c('words','nz',freq)} .}

\item{mode}{\code{insert} ,\code{get} or \code{remove} a word, dynamically updated dictionary.}
}
\value{
TRUE, FALSE or a character.
}
\description{
Dynamically updated dictionary
}
\examples{
\dontrun{
hanlp.updateWord('newword')
hanlp.updateWord(c('newword','nz',1000))
hanlp.updateWord(x=c('newword'),mode='get')
hanlp.updateWord(x=c('newword'),mode='remove')
}
}
\author{
\link{https://github.com/qxde01/RHanLP}
}
