% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hanlp.R
\name{hanlp.word2vecTrain}
\alias{hanlp.word2vecTrain}
\title{word2vec training.}
\usage{
hanlp.word2vecTrain(file_path = NULL, model_path = "word2vec",
  word_vectors_size = 128L, num_iters = 10L, word_min_freq = 5L,
  window_size = 5L, neg_sample = 5L, type = "skipgram",
  alpha = ifelse(type == "skipgram", 0.05, 0.025), softmax = TRUE,
  threads_num = 2L)
}
\arguments{
\item{file_path}{Corpus file path .}

\item{model_path}{Word2vec mode saving path.}

\item{word_vectors_size}{Word vector size ,default 128.}

\item{num_iters}{Number of iterations,default 10.}

\item{word_min_freq}{This will discard words that appear less than <int> times; default is 5}

\item{window_size}{Max skip length between words; default is 5.}

\item{neg_sample}{Negative sample number,default 5.}

\item{type}{\code{skipgram} or \code{cbow} model, default \code{skipgram}.}

\item{alpha}{the starting learning rate; default is 0.025 for \code{skipgram} and 0.05 for \code{cbow}.}

\item{threads_num}{the threads number default 2.}
}
\description{
A function of word2vec training .
}
\examples{
\dontrun{
hanlp.word2vecTrain(file_path='text.txt',model_path = 'word2vec.txt')
}
}
\author{
qxde01
}
