% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hanlp.R
\docType{data}
\name{hanlp.wordVectorModel}
\alias{hanlp.wordVectorModel}
\title{word Vector Model.}
\format{\code{\link{R6Class}} object.}
\usage{
hanlp.wordVectorModel
}
\description{
a R6 class of word Vector Model.
}
\details{
A R6class of wordVectorModel  .
}
\section{Usage}{

For usage details see \bold{Methods, Arguments and Examples} sections.
\preformatted{
wordVectorModel = hanlp.wordVectorModel$new(model_path)
wordVectorModel$similarity(word1,word2)
wordVectorModel$nearest(word,size=10L)
wordVectorModel$analogy(word1,word2,word3)
}
}

\section{Methods}{

\describe{
  \item{\code{$new(model_path)}}{Constructor for Global vectors model}
  \item{\code{$similarity(word1,word2)}}{Calculating the semantic distance of  \code{word1} and \code{word2}. }
  \item{\code{$nearest(word,size=10L)}}{Find out the \code{size}(default 10) words most similar to a \code{word}.}
  \item{\code{$analogy(word1,word2,word3)}}{Given three words:\code{word1},\code{word2},and \code{word3}, the words and their similarity
  list are returned to the semantic distance of (word1 - word2 + word3).}
}
}

\examples{
\dontrun{
wordvector<-hanlp.wordVectorModel$new(model_path)
wordvector$similarity('hello','world')
wordvector$nearest("world",size=10L)
wordvector$analogy('R','Python','Java')
}

}
\author{
qxde01
}
\keyword{datasets}
